Update examples to reflect real US cities
Reasoning and sources:

For lazy developers like me, I started using this package by copying the examples for the addresses because I figured that they were grounded in real-world, common validation rules. I did tweak the State and Zipcode validation to be more general, but then we hit a snag when we realized that Novi, MI (4 characters) is a real US city. This led me to look into the real boundaries and I've landed on this PR as an enhancement to save headaches for other future lazy developers. :)

- City between 1-50: Expand it to capture the set of real US cities. Source: 
  - https://www.serviceobjects.com/blog/character-limits-in-address-lines-for-usps-ups-and-fedex/ (no minimum, so I assume 1, maximum 50)

- State upper- or lower-case: I'm not strongly opinionated here, but I figured why be this restrictive in an example?

- Zipcode 5 or 9 digits: https://www.oreilly.com/library/view/regular-expressions-cookbook/9781449327453/ch04s14.html